[ALPS07450286] FPSGO: Use unsigned long long instead of int
Fix the timestamp of AA using type unsigned

MTK-Commit-Id: 2755c980ee5f3ccd904c0a1c237c0da118d2fca8

CR-Id: ALPS07450286
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v3/base/include/fpsgo_base.h
@@ -235,7 +235,7 @@ struct render_info {
 	int hwui;
 	int sbe_control_flag;
 	int control_pid_flag;
-	int render_last_cb_ts;
+	unsigned long long render_last_cb_ts;
 
 	/*render queue/dequeue/frame time info*/
 	unsigned long long t_enqueue_start;
